# api-eventos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Generación y Validación de documentos de participación",
        "description": "Listado de rutas de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/documents": {
            "get": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Validación del documento de participación digital.",
                "operationId": "e41cb858af257bb22df5de998012658b",
                "parameters": [
                    {
                        "name": "cadena",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Nombres": {
                                            "type": "string",
                                            "example": "Luis Mario"
                                        },
                                        "ApellidoPaterno": {
                                            "type": "string",
                                            "example": "Pérez"
                                        },
                                        "ApellidoMaterno": {
                                            "type": "string",
                                            "example": "Ramírez"
                                        },
                                        "Evento": {
                                            "type": "string",
                                            "example": "Academia Journals"
                                        },
                                        "Descripcion": {
                                            "type": "text",
                                            "example": "Espacio de trabajo que da a conocer los resultados de trabajos de investigación."
                                        },
                                        "TipoParticipante": {
                                            "type": "string",
                                            "example": "Ponente"
                                        },
                                        "FechaTermino": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-05-29 17:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos no coinciden"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Almacena un recurso creado - genera documentos de participación.",
                "operationId": "9b247d60e8834d470666fba8e773edb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "Nombres",
                                    "ApellidoPaterno",
                                    "ApellidoMaterno",
                                    "Correo",
                                    "Evento",
                                    "Descripcion",
                                    "TipoParticipante",
                                    "FechaTermino"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "Nombres": {
                                        "type": "string",
                                        "example": "Luis Mario"
                                    },
                                    "ApellidoPaterno": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "ApellidoMaterno": {
                                        "type": "string",
                                        "example": "Ramírez"
                                    },
                                    "Correo": {
                                        "type": "string",
                                        "example": "vapire117@gmail.com"
                                    },
                                    "Evento": {
                                        "type": "string",
                                        "example": "Academia Journals"
                                    },
                                    "Descripcion": {
                                        "type": "text",
                                        "example": "Espacio de trabajo que da a conocer los resultados de trabajos de investigación."
                                    },
                                    "TipoParticipante": {
                                        "type": "string",
                                        "example": "Ponente"
                                    },
                                    "FechaTermino": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-29 17:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Documentos generados y enviados"
                                        },
                                        "icono": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Faltan campos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}